commit when last commit was too long ago
When indexing mail with unusually long headers, indexing becomes slow
and memory hungry. Indexing such emails by batch of 1000 ends up eating
all the memory, and the index process is killed. The next time indexing
is triggered, all the work is lost so no progress can be done.

This patch enforces a commit every 5 minutes. This ensures that even
though only 36 emails (real number) are indexed in the meantime, and the
process is killed afterwards because memory usage is high, progress can
be done.

About the choice of 5 minutes: commit time will be negligible when this
triggers.  In my workload, I saw both cases where the number of
update and the time were the limiting factor triggering the commit.